+++ src/SpotifyWebAPI.php
@@ -81,7 +81,6 @@ class SpotifyWebAPI
 
     /**
      * Add albums to the current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/save-albums-user/
      *
      * @param string|array $albums ID(s) of the album(s) to add.
@@ -105,7 +104,6 @@ class SpotifyWebAPI
 
     /**
      * Add tracks to the current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/save-tracks-user/
      *
      * @param string|array $tracks ID(s) of the track(s) to add.
@@ -129,7 +127,6 @@ class SpotifyWebAPI
 
     /**
      * Add tracks to a user's playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/add-tracks-to-playlist/
      *
      * @param string $userId ID of the user who owns the playlist.
@@ -159,14 +156,63 @@ class SpotifyWebAPI
     }
 
     /**
+     * Change the current user's playback device.
+     * https://developer.spotify.com/web-api/transfer-a-users-playback/
+     *
+     * @param array|object $options Options for the playback transfer.
+     * - string|array device_ids Required. ID of the device to switch to.
+     * - bool play Optional. Whether to start playing on the new device
+     *
+     * @return bool Whether the playback device was successfully changed.
+     */
+    public function changeMyDevice($options)
+    {
+        $options = (array) $options;
+        $options['device_ids'] = (array) $options['device_ids'];
+        $options = json_encode($options);
+
+        $headers = $this->authHeaders();
+        $headers['Content-Type'] = 'application/json';
+
+        $uri = '/v1/me/player';
+
+        $this->lastResponse = $this->request->api('PUT', $uri, $options, $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
+     * Change playback volume for the current user.
+     * https://developer.spotify.com/web-api/set-volume-for-users-playback/
+     *
+     * @param array|object $options Optional. Options for the playback volume.
+     * - int volume_percent Required. The volume to set.
+     * - string device_id Optional. ID of the device to target.
+     *
+     * @return bool Whether the playback volume was successfully changed.
+     */
+    public function changeVolume($options)
+    {
+        $options = http_build_query($options);
+
+        $headers = $this->authHeaders();
+
+        // We need to manually append data to the URI since it's a PUT request
+        $uri = '/v1/me/player/volume?' . $options;
+
+        $this->lastResponse = $this->request->api('PUT', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
      * Create a new playlist for a user.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/create-playlist/
      *
      * @param string $userId ID of the user to create the playlist for.
      * @param array|object $options Options for the new playlist.
-     * - name string Required. Name of the playlist.
-     * - public bool Optional. Whether the playlist should be public or not.
+     * - string name Required. Name of the playlist.
+     * - bool public Optional. Whether the playlist should be public or not.
      *
      * @return array|object The new playlist. Type is controlled by `SpotifyWebAPI::setReturnType()`.
      */
@@ -186,7 +232,6 @@ class SpotifyWebAPI
 
     /**
      * Check to see if the current user is following one or more artists or other Spotify users.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/check-current-user-follows/
      *
      * @param string $type The type to check: either 'artist' or 'user'.
@@ -213,7 +258,6 @@ class SpotifyWebAPI
 
     /**
      * Delete albums from current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/remove-albums-user/
      *
      * @param string|array $albums ID(s) of the album(s) to delete.
@@ -223,9 +267,7 @@ class SpotifyWebAPI
     public function deleteMyAlbums($albums)
     {
         $albums = $this->uriToId($albums, 'album');
-        $albums = json_encode(
-            (array) $albums
-        );
+        $albums = json_encode((array) $albums);
 
         $headers = $this->authHeaders();
         $headers['Content-Type'] = 'application/json';
@@ -239,7 +281,6 @@ class SpotifyWebAPI
 
     /**
      * Delete tracks from current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/remove-tracks-user/
      *
      * @param string|array $tracks ID(s) of the track(s) to delete.
@@ -249,9 +290,7 @@ class SpotifyWebAPI
     public function deleteMyTracks($tracks)
     {
         $tracks = $this->uriToId($tracks);
-        $tracks = json_encode(
-            (array) $tracks
-        );
+        $tracks = json_encode((array) $tracks);
 
         $headers = $this->authHeaders();
         $headers['Content-Type'] = 'application/json';
@@ -265,14 +304,13 @@ class SpotifyWebAPI
 
     /**
      * Delete tracks from a playlist and retrieve a new snapshot ID.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/remove-tracks-playlist/
      *
      * @param string $userId ID of the user who owns the playlist.
      * @param string $playlistId ID of the playlist to delete tracks from.
      * @param array $tracks Array of arrays or objects with tracks to delete.
-     * - id string Required. Spotify track ID.
-     * - positions int|array Optional. The track's position(s) in the playlist.
+     * - string id Required. Spotify track ID.
+     * - int|array positions Optional. The track's position(s) in the playlist.
      * @param string $snapshotId Optional. The playlist's snapshot ID.
      *
      * @return string|bool A new snapshot ID or false if the tracks weren't successfully deleted.
@@ -318,7 +356,6 @@ class SpotifyWebAPI
 
     /**
      * Add the current user as a follower of one or more artists or other Spotify users.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/follow-artists-users/
      *
      * @param string $type The type to check: either 'artist' or 'user'.
@@ -345,19 +382,18 @@ class SpotifyWebAPI
 
     /**
      * Add the current user as a follower of a playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/follow-playlist/
      *
      * @param string $userId ID of the user who owns the playlist.
      * @param string $playlistId ID of the playlist to follow.
      * @param array|object $options Optional. Options for the followed playlist.
-     * - public bool Optional. Whether the playlist should be followed publicly or not.
+     * - bool public Optional. Whether the playlist should be followed publicly or not.
      *
      * @return bool Whether the playlist was successfully followed.
      */
     public function followPlaylist($userId, $playlistId, $options = [])
     {
-        $options = json_encode($options);
+        $options = json_encode((object) $options);
 
         $headers = $this->authHeaders();
         $headers['Content-Type'] = 'application/json';
@@ -550,7 +586,6 @@ class SpotifyWebAPI
 
     /**
      * Get track audio features.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-several-audio-features/
      *
      * @param array $trackIds IDs of the tracks.
@@ -573,8 +608,26 @@ class SpotifyWebAPI
     }
 
     /**
+     * Get audio analysis for track.
+     * https://developer.spotify.com/web-api/get-audio-analysis/
+     *
+     * @param string $trackId ID of the track.
+     *
+     * @return object The track's audio analysis. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+     */
+    public function getAudioAnalysis($trackId)
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/audio-analysis/' . $trackId;
+
+        $this->lastResponse = $this->request->api('GET', $uri, [], $headers);
+
+        return $this->lastResponse['body'];
+    }
+
+    /**
      * Get a list of categories used to tag items in Spotify (on, for example, the Spotify player’s "Browse" tab).
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-list-categories/
      *
      * @param array|object $options Optional. Options for the categories.
@@ -598,7 +651,6 @@ class SpotifyWebAPI
 
     /**
      * Get a single category used to tag items in Spotify (on, for example, the Spotify player’s "Browse" tab).
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-category/
      *
      * @param string $categoryId The Spotify ID of the category.
@@ -622,7 +674,6 @@ class SpotifyWebAPI
 
     /**
      * Get a list of Spotify playlists tagged with a particular category.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-categorys-playlists/
      *
      * @param string $categoryId The Spotify ID of the category.
@@ -647,7 +698,6 @@ class SpotifyWebAPI
 
     /**
      * Get Spotify featured playlists.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-list-featured-playlists/
      *
      * @param array|object $options Optional. Options for the playlists.
@@ -672,7 +722,6 @@ class SpotifyWebAPI
 
     /**
      * Get a list of possible seed genres.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-recommendations/#available-genre-seeds
      *
      * @return array|object All possible seed genres. Type is controlled by `SpotifyWebAPI::setReturnType()`.
@@ -703,22 +752,19 @@ class SpotifyWebAPI
     }
 
     /**
-     * Get new releases.
-     * Requires a valid access token.
-     * https://developer.spotify.com/web-api/get-list-new-releases/
+     * Get the current user’s currently playing track.
+     * https://developer.spotify.com/web-api/get-the-users-currently-playing-track/
      *
-     * @param array|object $options Optional. Options for the items.
-     * - string country Optional. An ISO 3166-1 alpha-2 country code. Show items relevant to this country.
-     * - int limit Optional. Limit the number of items.
-     * - int offset Optional. Number of items to skip.
+     * @param array|object $options Optional. Options for the track.
+     * - string market Optional. An ISO 3166-1 alpha-2 country code, provide this if you wish to apply Track Relinking.
      *
-     * @return array|object The new releases. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+     * @return array|object The user's currently playing track. Type is controlled by `SpotifyWebAPI::setReturnType()`.
      */
-    public function getNewReleases($options = [])
+    public function getMyCurrentTrack($options = [])
     {
         $headers = $this->authHeaders();
 
-        $uri = '/v1/browse/new-releases';
+        $uri = '/v1/me/player/currently-playing';
 
         $this->lastResponse = $this->request->api('GET', $uri, $options, $headers);
 
@@ -726,8 +772,45 @@ class SpotifyWebAPI
     }
 
     /**
+     * Get the current user’s devices.
+     * https://developer.spotify.com/web-api/get-a-users-available-devices/
+     *
+     * @return array|object The user's devices. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+     */
+    public function getMyDevices()
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player/devices';
+
+        $this->lastResponse = $this->request->api('GET', $uri, [], $headers);
+
+        return $this->lastResponse['body'];
+    }
+
+    /**
+     * Get the current user’s current playback information.
+     * https://developer.spotify.com/web-api/get-information-about-the-users-current-playback/
+     *
+     * @param array|object $options Optional. Options for the info.
+     * - string market Optional. An ISO 3166-1 alpha-2 country code, provide this if you wish to apply Track Relinking.
+     *
+     * @return array|object The user's playback information. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+     */
+    public function getMyCurrentPlaybackInfo($options = [])
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player';
+
+        $this->lastResponse = $this->request->api('GET', $uri, $options, $headers);
+
+        return $this->lastResponse['body'];
+    }
+
+
+    /**
      * Get the current user’s playlists.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-a-list-of-current-users-playlists/
      *
      * @param array|object $options Optional. Options for the playlists.
@@ -748,12 +831,35 @@ class SpotifyWebAPI
     }
 
     /**
+      * Get the current user’s recently played tracks.
+      * https://developer.spotify.com/web-api/web-api-personalization-endpoints/get-recently-played/
+      *
+      * @param array|object $options Optional. Options for the tracks.
+      * - int limit Optional. Number of tracks to return.
+      * - string after Optional. Unix timestamp in ms (13 digits). Returns all items after this position.
+      * - string before Optional. Unix timestamp in ms (13 digits). Returns all items before this position.
+      *
+      * @return array|object The most recently played tracks. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+      */
+    public function getMyRecentTracks($options = [])
+    {
+        $options = (array) $options;
+
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player/recently-played';
+
+        $this->lastResponse = $this->request->api('GET', $uri, $options, $headers);
+
+        return $this->lastResponse['body'];
+    }
+
+    /**
      * Get the current user’s saved albums.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-users-saved-albums/
      *
      * @param array|object $options Optional. Options for the albums.
-     * - int limit Optional. Limit the number of albums.
+     * - int limit Optional. Number of albums to return.
      * - int offset Optional. Number of albums to skip.
      * - string market Optional. An ISO 3166-1 alpha-2 country code, provide this if you wish to apply Track Relinking.
      *
@@ -772,7 +878,6 @@ class SpotifyWebAPI
 
     /**
      * Get the current user’s saved tracks.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-users-saved-tracks/
      *
      * @param array|object $options Optional. Options for the tracks.
@@ -795,7 +900,6 @@ class SpotifyWebAPI
 
     /**
      * Get the current user's top tracks or artists.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-users-top-artists-and-tracks/
      *
      * @param string $type The type of entity to fetch.
@@ -818,8 +922,29 @@ class SpotifyWebAPI
     }
 
     /**
+     * Get new releases.
+     * https://developer.spotify.com/web-api/get-list-new-releases/
+     *
+     * @param array|object $options Optional. Options for the items.
+     * - string country Optional. An ISO 3166-1 alpha-2 country code. Show items relevant to this country.
+     * - int limit Optional. Limit the number of items.
+     * - int offset Optional. Number of items to skip.
+     *
+     * @return array|object The new releases. Type is controlled by `SpotifyWebAPI::setReturnType()`.
+     */
+    public function getNewReleases($options = [])
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/browse/new-releases';
+
+        $this->lastResponse = $this->request->api('GET', $uri, $options, $headers);
+
+        return $this->lastResponse['body'];
+    }
+
+    /**
      * Get recommendations based on artists, tracks, or genres.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-recommendations/
      *
      * @param array|object $options Optional. Options for the recommendations.
@@ -955,7 +1080,6 @@ class SpotifyWebAPI
 
     /**
      * Get the artists followed by the current user.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-followed-artists/
      *
      * @param array|object $options Optional. Options for the artists.
@@ -983,7 +1107,6 @@ class SpotifyWebAPI
 
     /**
      * Get a user's specific playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-playlist/
      *
      * @param string $userId ID of the user.
@@ -1013,7 +1136,6 @@ class SpotifyWebAPI
 
     /**
      * Get a user's playlists.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-list-users-playlists/
      *
      * @param string $userId ID of the user.
@@ -1036,7 +1158,6 @@ class SpotifyWebAPI
 
     /**
      * Get the tracks in a user's playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-playlists-tracks/
      *
      * @param string $userId ID of the user.
@@ -1068,7 +1189,6 @@ class SpotifyWebAPI
 
     /**
      * Get the currently authenticated user.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/get-current-users-profile/
      *
      * @return array|object The currently authenticated user. Type is controlled by `SpotifyWebAPI::setReturnType()`.
@@ -1086,7 +1206,6 @@ class SpotifyWebAPI
 
     /**
      * Check if albums are saved in the current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/check-users-saved-albums/
      *
      * @param string|array $albums ID(s) of the album(s) to check for.
@@ -1095,7 +1214,9 @@ class SpotifyWebAPI
      */
     public function myAlbumsContains($albums)
     {
+        $albums = $this->uriToId($albums, 'album');
         $albums = implode(',', (array) $albums);
+
         $options = [
             'ids' => $albums,
         ];
@@ -1111,7 +1232,6 @@ class SpotifyWebAPI
 
     /**
      * Check if tracks are saved in the current user's Spotify library.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/check-users-saved-tracks/
      *
      * @param string|array $tracks ID(s) of the track(s) to check for.
@@ -1120,7 +1240,9 @@ class SpotifyWebAPI
      */
     public function myTracksContains($tracks)
     {
+        $tracks = $this->uriToId($tracks);
         $tracks = implode(',', (array) $tracks);
+
         $options = [
             'ids' => $tracks,
         ];
@@ -1135,8 +1257,110 @@ class SpotifyWebAPI
     }
 
     /**
+     * Play the next track in the current users's queue.
+     * https://developer.spotify.com/web-api/skip-users-playback-to-next-track/
+     *
+     * @param string $deviceId Optional. ID of the device to target.
+     *
+     * @return bool Whether the track was successfully skipped.
+     */
+    public function next($deviceId = '')
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player/next';
+
+        // We need to manually append data to the URI since it's a POST request
+        if ($deviceId) {
+            $uri = $uri . '?device_id=' . $deviceId;
+        }
+
+        $this->lastResponse = $this->request->api('POST', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
+     * Pause playback for the current user.
+     * https://developer.spotify.com/web-api/pause-a-users-playback/
+     *
+     * @param string $deviceId Optional. ID of the device to pause on.
+     *
+     * @return bool Whether the playback was successfully paused.
+     */
+    public function pause($deviceId = '')
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player/pause';
+
+        // We need to manually append data to the URI since it's a PUT request
+        if ($deviceId) {
+            $uri = $uri . '?device_id=' . $deviceId;
+        }
+
+        $this->lastResponse = $this->request->api('PUT', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
+     * Start playback for the current user.
+     * https://developer.spotify.com/web-api/start-a-users-playback/
+     *
+     * @param string $deviceId Optional. ID of the device to play on.
+     * @param array|object $options Optional. Options for the playback.
+     * - string context_uri Optional. Spotify URI of the context to play, for example an album.
+     * - array uris Optional. Spotify track URIs to play.
+     * - object offset Optional. Indicates from where in the context playback should start.
+     *
+     * @return bool Whether the playback was successfully started.
+     */
+    public function play($deviceId = '', $options = [])
+    {
+        $options = json_encode((object) $options);
+
+        $headers = $this->authHeaders();
+        $headers['Content-Type'] = 'application/json';
+
+        $uri = '/v1/me/player/play';
+
+        // We need to manually append data to the URI since it's a PUT request
+        if ($deviceId) {
+            $uri = $uri . '?device_id=' . $deviceId;
+        }
+
+        $this->lastResponse = $this->request->api('PUT', $uri, $options, $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
+     * Play the previous track in the current users's queue.
+     * https://developer.spotify.com/web-api/skip-users-playback-to-previous-track/
+     *
+     * @param string $deviceId Optional. ID of the device to target.
+     *
+     * @return bool Whether the track was successfully skipped.
+     */
+    public function previous($deviceId = '')
+    {
+        $headers = $this->authHeaders();
+
+        $uri = '/v1/me/player/previous';
+
+        // We need to manually append data to the URI since it's a POST request
+        if ($deviceId) {
+            $uri = $uri . '?device_id=' . $deviceId;
+        }
+
+        $this->lastResponse = $this->request->api('POST', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
      * Reorder the tracks in a user's playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/reorder-playlists-tracks/
      *
      * @param string $userId ID of the user.
@@ -1169,8 +1393,31 @@ class SpotifyWebAPI
     }
 
     /**
+     * Set repeat mode for the current user’s playback.
+     * https://developer.spotify.com/web-api/set-repeat-mode-on-users-playback/
+     *
+     * @param array|object $options Optional. Options for the playback repeat mode.
+     * - string state Required. The repeat mode. See Spotify docs for possible values.
+     * - string device_id Optional. ID of the device to target.
+     *
+     * @return bool Whether the playback repeat mode was successfully changed.
+     */
+    public function repeat($options)
+    {
+        $options = http_build_query($options);
+
+        $headers = $this->authHeaders();
+
+        // We need to manually append data to the URI since it's a PUT request
+        $uri = '/v1/me/player/repeat?' . $options;
+
+        $this->lastResponse = $this->request->api('PUT', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
      * Replace all tracks in a user's playlist with new ones.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/replace-playlists-tracks/
      *
      * @param string $userId ID of the user.
@@ -1228,6 +1475,30 @@ class SpotifyWebAPI
     }
 
     /**
+     * Change playback position for the current user.
+     * https://developer.spotify.com/web-api/seek-to-position-in-currently-playing-track/
+     *
+     * @param array|object $options Optional. Options for the playback seeking.
+     * - string position_ms Required. The position in milliseconds to seek to.
+     * - string device_id Optional. ID of the device to target.
+     *
+     * @return bool Whether the playback position was successfully changed.
+     */
+    public function seek($options)
+    {
+        $options = http_build_query($options);
+
+        $headers = $this->authHeaders();
+
+        // We need to manually append data to the URI since it's a PUT request
+        $uri = '/v1/me/player/seek?' . $options;
+
+        $this->lastResponse = $this->request->api('PUT', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
      * Set the access token to use.
      *
      * @param string $accessToken The access token.
@@ -1273,8 +1544,33 @@ class SpotifyWebAPI
     }
 
     /**
+     * Set shuffle mode for the current user’s playback.
+     * https://developer.spotify.com/web-api/toggle-shuffle-for-users-playback/
+     *
+     * @param array|object $options Optional. Options for the playback shuffle mode.
+     * - bool state Required. The shuffle mode. See Spotify docs for possible values.
+     * - string device_id Optional. ID of the device to target.
+     *
+     * @return bool Whether the playback shuffle mode was successfully changed.
+     */
+    public function shuffle($options)
+    {
+        $options = (array) $options;
+        $options['state'] = $options['state'] ? 'true' : 'false';
+        $options = http_build_query($options);
+
+        $headers = $this->authHeaders();
+
+        // We need to manually append data to the URI since it's a PUT request
+        $uri = '/v1/me/player/shuffle?' . $options;
+
+        $this->lastResponse = $this->request->api('PUT', $uri, [], $headers);
+
+        return $this->lastResponse['status'] == 204;
+    }
+
+    /**
      * Remove the current user as a follower of one or more artists or other Spotify users.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/unfollow-artists-users/
      *
      * @param string $type The type to check: either 'artist' or 'user'.
@@ -1301,7 +1597,6 @@ class SpotifyWebAPI
 
     /**
      * Remove the current user as a follower of a playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/unfollow-playlist/
      *
      * @param string $userId ID of the user who owns the playlist.
@@ -1323,7 +1618,6 @@ class SpotifyWebAPI
 
     /**
      * Update the details of a user's playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/change-playlist-details/
      *
      * @param string $userId ID of the user who owns the playlist.
@@ -1350,7 +1644,6 @@ class SpotifyWebAPI
 
     /**
      * Check if a user is following a playlist.
-     * Requires a valid access token.
      * https://developer.spotify.com/web-api/check-user-following-playlist/
      *
      * @param string $ownerId User ID of the playlist owner.